projects/inoah-dating/requirements.txt
# iNoah Dating Dependencies
fastapi>=0.100.0
uvicorn>=0.22.0
httpx>=0.24.0
Pillow>=9.0.0

# Note: Also requires inoah-core to be installed
# pip install -e ../inoah-core
